Separando comandos em módulos diferentes.
